backend/requirements.txt
# LLM integrations - versión más reciente compatible
google-genai

# AWS
boto3==1.35.36

# Utilities
requests==2.31.0
pydantic==2.8.2

# Development
pytest==8.3.3
black==24.8.0